fix(scan): fix architecture-gate violations from the abi3 dry-run check
CI's ai-readiness job caught two real issues in the previous commit:

- debt-no-growth: cli_scan.py and service_scan.py are pinned at their
  1993-line adoption baseline, and the abi3/query-only additions pushed
  both over it (they were already at the edge, under the AI-readiness
  2000-line hard cap but not the stricter per-file no-growth debt).
- dependency-direction: cli_scan.py and artifact_set_dry_run.py are
  `frontends`-classified (ADR-061) and may not import `python_ext.py`
  (`extract`) directly -- only `model`/`workflows`/`report`.

Fixed by moving the glue into a new `workflows`-layer module
(abicheck/workflows/scan_abi3_dry_run.py, which depends only on
python_ext and so sits outside the CLI-registration import cycle), and
splitting render_scan_dry_run out of cli_scan.py into
abicheck/frontends/cli/scan_dry_run.py -- the same split its
--artifact-set sibling (artifact_set_dry_run.py) already went through
one PR earlier, for the identical reason. service_scan.py's TU-estimate
change is now a net-zero-line edit (a single-line conditional replacing
the previous fixed string) so it doesn't need a similar extraction.

Verified: scripts/check_architecture.py now reports 0 errors (was 6),
mypy/ruff clean, and the full targeted test suite (419 tests across the
touched modules, including test_cli_contract.py's Tier-1 boundary gate)
passes.
